semantic_rag_answer/requirements.txt
google-cloud-bigquery
weaviate-client>=4.3.0
sentence-transformers
flask
db-dtypes
together
